ray-tracing2.py now uses ADParser.py as a module to do all of its input from the art description file. This led to the discovery of a bug in ADParser.py: the components of the resolution should be integers, not floating points, so they need to be interpreted that way when parsing the file.

+++ ADParser.py
@@ -39,7 +39,7 @@ def parse_resolution(file):
   parts = line.split('=')
   if len(parts) != 2 or parts[0] != 'resolution':
     raise ADParseError
-  resolution = [float(x) for x in parts[1].split(',')]
+  resolution = [int(x) for x in parts[1].split(',')]
   if len(resolution) != 2:
     raise ADParseError
   return resolution
